Move streamResponseToUser and replacePlaceholdersDuringStreaming to angelaUtils
